+++ src/saxpy/plain_kernel.h
@@ -1,6 +1,6 @@
 #ifndef PLAIN_KERNEL_H
 #define PLAIN_KERNEL_H
 
-void saxpy_plain(const int *x, int *y, const int a, const int arr_size);
+void saxpy_plain(const float *x, float *y, const float a, const int arr_size);
 
 #endif // PLAIN_KERNEL_H
